requirements.txt
evdev==1.6.1
pynput==1.7.6
python-dotenv==1.0.0
python-xlib==0.33
six==1.16.0
